#56 refactor : AnimatedBubble 파일 분리
